// cypress/support/pageObjects/dataEntryForm/enums/dynamicWebTableEnum.ts
export enum BrowserNames {
    FIREFOX = "Firefox",
    CHROME = "Chrome",
    SYSTEM = "System",
    INTERNET_EXPLORER = "Internet Explorer"
}


export enum DynamicWebTableHeadersEnum {
    NAME = "Name",
    DISK = "Disk (MB/s)",
    NETWORK = "Network (Mbps)",
    MEMORY = "Memory (MB)",
    CPU = "CPU (%)"
}